core_first_steps: Move content from netdata_ebpf to netdata_common, because they will be necessary for different kind of eBPF programs
